# Conflicts:
#	src/main/java/com/utn/EBS/Controladores/PedidoController.java
#	src/main/java/com/utn/EBS/Repositorios/PedidoRepository.java
#	src/main/java/com/utn/EBS/Servicios/PedidoServiceImpl.java

+++ src/main/java/com/utn/EBS/Entidades/Ingrediente.java
@@ -33,7 +33,7 @@ public class Ingrediente extends BaseEntidad{
     private String nombre;
 
     @Column(name = "costo", nullable = false)
-    private float costo;
+    private double costo;
 
     @Column(name = "stock_actual", nullable = false)
     private int stockActual;
